+++ packages/template-retail-react-app/app/pages/cart/index.jsx
@@ -82,7 +82,7 @@ const Cart = () => {
                     // we must use `formatMessage()`, rather than `<FormattedMessage />`.
                     <Button variant="link" onClick={() => navigate('/account/wishlist')}>
                         {formatMessage({
-                            id: 'cart.link.added_to_wishlist.view_wishlist',
+                            id: 'wishlist.link.added_to_wishlist.view',
                             defaultMessage: 'View'
                         })}
                     </Button>