+++ packages/template-retail-react-app/app/pages/product-detail/index.jsx
@@ -100,7 +100,7 @@ const ProductDetail = ({category, product, isLoading}) => {
                     // we must use `formatMessage()`, rather than `<FormattedMessage />`.
                     <Button variant="link" onClick={() => navigate('/account/wishlist')}>
                         {formatMessage({
-                            id: 'cart.link.added_to_wishlist.view_wishlist',
+                            id: 'wishlist.link.added_to_wishlist.view',
                             defaultMessage: 'View'
                         })}
                     </Button>