+++ packages/template-retail-react-app/app/pages/product-list/index.jsx
@@ -146,7 +146,7 @@ const ProductList = (props) => {
                     // we must use `formatMessage()`, rather than `<FormattedMessage />`.
                     <Button variant="link" onClick={() => navigate('/account/wishlist')}>
                         {formatMessage({
-                            id: 'cart.link.added_to_wishlist.view_wishlist',
+                            id: 'wishlist.link.added_to_wishlist.view',
                             defaultMessage: 'View'
                         })}
                     </Button>